:sparkles: Skip brew

+++ tools/dump.sh
@@ -1,8 +1,19 @@
 #!/bin/bash
 
+while getopts b OPT
+do
+  case $OPT in
+    b) BREW=1
+      ;;
+  esac
+done
+
 cp ~/.zshrc .
 cp ~/.gitconfig .
 
-rm ./Brewfile
-rm ./Brewfile.lock.json
-brew bundle dump
+if [[ $BREW ]]; then
+  rm ./Brewfile
+  rm ./Brewfile.lock.json
+  brew bundle dump
+fi
+
